requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
requests>=2.25.0
